+++ sdk/account-info/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "solana-account-info"
+description = "Solana AccountInfo and related definitions."
+documentation = "https://docs.rs/solana-account-info"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[dependencies]
+bincode = { workspace = true }
+serde = { workspace = true }
+solana-program-error = { workspace = true }
+solana-program-memory = { workspace = true }
+solana-pubkey = { workspace = true, default-features = false }
+
+[dev-dependencies]
+solana-program = { workspace = true }
+static_assertions = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
