+++ sdk/program/Cargo.toml
@@ -31,6 +31,7 @@ serde_bytes = { workspace = true }
 serde_derive = { workspace = true }
 sha2 = { workspace = true }
 sha3 = { workspace = true }
+solana-account-info = { workspace = true }
 solana-atomic-u64 = { workspace = true }
 solana-clock = { workspace = true, features = ["serde"] }
 solana-decode-error = { workspace = true }
