+++ sdk/program/src/lib.rs
@@ -469,7 +469,6 @@
 // Allows macro expansion of `use ::solana_program::*` to work within this crate
 extern crate self as solana_program;
 
-pub mod account_info;
 pub mod address_lookup_table;
 pub mod big_mod_exp;
 pub mod blake3;
@@ -483,7 +482,6 @@ pub mod bpf_loader;
 pub mod bpf_loader_deprecated;
 pub mod bpf_loader_upgradeable;
 pub mod compute_units;
-pub mod debug_account_data;
 pub mod ed25519_program;
 pub mod entrypoint;
 pub mod entrypoint_deprecated;
@@ -539,6 +537,7 @@ pub use solana_short_vec as short_vec;
 #[cfg(target_arch = "wasm32")]
 pub use wasm_bindgen::prelude::wasm_bindgen;
 pub use {
+    solana_account_info::{self as account_info, debug_account_data},
     solana_clock as clock, solana_msg::msg, solana_program_option as program_option,
     solana_pubkey as pubkey,
 };
